+++ src/main/java/com/octagram/pollet/auth/domain/status/AuthSuccessCode.java
@@ -2,7 +2,7 @@ package com.octagram.pollet.auth.domain.status;
 
 import org.springframework.http.HttpStatus;
 
-import com.octagram.pollet.global.status.BaseCode;
+import com.octagram.pollet.global.domain.status.BaseCode;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
