+++ src/main/java/com/octagram/pollet/gifticon/domain/status/GifticonErrorCode.java
@@ -2,7 +2,7 @@ package com.octagram.pollet.gifticon.domain.status;
 
 import org.springframework.http.HttpStatus;
 
-import com.octagram.pollet.global.status.BaseCode;
+import com.octagram.pollet.global.domain.status.BaseCode;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
