+++ src/main/java/com/octagram/pollet/global/advisor/ApiResponseAdvice.java
@@ -8,7 +8,7 @@ import org.springframework.http.server.ServerHttpResponse;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
 
-import com.octagram.pollet.global.dto.ApiResponse;
+import com.octagram.pollet.global.presentation.dto.ApiResponse;
 
 import lombok.NonNull;
