+++ src/main/java/com/octagram/pollet/member/domain/status/MemberSuccessCode.java
@@ -2,7 +2,7 @@ package com.octagram.pollet.member.domain.status;
 
 import org.springframework.http.HttpStatus;
 
-import com.octagram.pollet.global.status.BaseCode;
+import com.octagram.pollet.global.domain.status.BaseCode;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
