+++ src/main/java/com/octagram/pollet/survey/domain/status/SurveySuccessCode.java
@@ -1,6 +1,6 @@
 package com.octagram.pollet.survey.domain.status;
 
-import com.octagram.pollet.global.status.BaseCode;
+import com.octagram.pollet.global.domain.status.BaseCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
